feat: added layer into an existing best implementaion
