+++ src/layers/controllers/layersController.ts
@@ -2,11 +2,13 @@ import { RequestHandler } from 'express';
 import httpStatus from 'http-status-codes';
 import { injectable, inject } from 'tsyringe';
 import { Services } from '../../common/constants';
-import { ILogger, ILayerPostRequest } from '../../common/interfaces';
+import { ILogger, ILayerPostRequest, ILayerToBestRequest, IBestLayer } from '../../common/interfaces';
 import { LayersManager } from '../models/layersManager';
 
 type CreateLayerHandler = RequestHandler<undefined, ILayerPostRequest, ILayerPostRequest>;
 type GetLayerHandler = RequestHandler<undefined, ILayerPostRequest>;
+type CreateBestHandler = RequestHandler<undefined, ILayerToBestRequest, ILayerToBestRequest>;
+type GetBestHandler = RequestHandler<undefined, IBestLayer>;
 
 @injectable()
 export class LayersController {
@@ -19,6 +21,15 @@ export class LayersController {
   public addLayer: CreateLayerHandler = (req, res, next) => {
     try {
       this.manager.addLayer(req.body);
+      return res.status(httpStatus.CREATED);
+    } catch (error) {
+      next(error);
+    }
+  };
+
+  public addLayerToBest: CreateBestHandler = (req, res, next) => {
+    try {
+      this.manager.addLayerToBest(req.body);
       return res.status(httpStatus.CREATED);
     } catch (error) {
       next(error);
